+++ gfx/framework/OtUi.h
@@ -185,8 +185,8 @@ public:
 	// knobs
 	static bool knob(const char* label, float* value, float minValue, float maxValue, const char* format="%.0f", bool logarithmic=false);
 	static bool knob(const char* label, int* value, int minValue, int maxValue);
-	static float knobWidth(int columns=1);
-	static float knobHeight(int rows=1);
+	static float knobWidth(size_t columns=1);
+	static float knobHeight(size_t rows=1);
 
 	static bool trimSlider(float* value);
 	static float trimSliderWidth();
